+++ Fahrtenbuch.Web/Constants.cs
@@ -1,0 +1,8 @@
+﻿namespace Fahrtenbuch.Web
+{
+    public class Constants
+    {
+        public const string LoginPath = "/Account/Login";
+        public const string LogoutPath = "/Account/Logout";
+    }
+}
