+++ Fahrtenbuch/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace Fahrtenbuch
+{
+    public class Constants
+    {
+        public const string Auth0Claim = "Auth0";
+    }
+}
